feat(plugin owner): adds logic
