+++ lib/version.rb
@@ -1,1 +1,1 @@
-VERSION = "0.60.0"
+VERSION = "0.61.0"
